// src/client/AppData.ts
import {Baby} from "./models/Baby";
import {BabyDateOfBirth} from "./models/BabyDateOfBirth";
import {BabyImage} from "./models/BabyImage";
import {ExcretionNotificationTrigger} from "./models/ExcretionNotificationTrigger";
import {Pumping} from "./models/Pumping";
import {PumpingNotificationTrigger} from "./models/PumpingNotificationTrigger";
import {PumpingTimeOfDayNotificationTrigger} from "./models/PumpingTimeOfDayNotificationTrigger";
import {Sleep} from "./models/Sleep";
import {SleepingNotificationTrigger} from "./models/SleepingNotificationTrigger";
import {Teeth} from "./models/Teeth";
import {Temperature} from "./models/Temperature";
import {Vaccination} from "./models/Vaccination";
import {JournalEntry} from "./models/JournalEntry";
import {Feed} from "./models/Feed";
import {Excretion} from "./models/Excretion";
import {Medicine} from "./models/Medicine";
import {Pause} from "./models/Pause";
import {Growth} from "./models/Growth";
import {SleepingsTimeOfDayNotificationTrigger} from "./models/SleepingsTimeOfDayNotificationTrigger";
import {MedicineRecord} from "./models/MedicineRecord";
import {ExcretionTimeOfDayNotificationTrigger} from "./models/ExcretionTimeOfDayNotificationTrigger";
import {FeedingNotificationTrigger} from "./models/FeedingNotificationTrigger";
import {FeedingTimeOfDayNotificationTrigger} from "./models/FeedingTimeOfDayNotificationTrigger";
import {GeneralNotesNotificationTrigger} from "./models/GeneralNotesNotificationTrigger";
import {GeneralNotesTimeOfDayNotificationTrigger} from "./models/GeneralNotesTimeOfDayNotificationTrigger";
import {MedicationRecordsNotificationTrigger} from "./models/MedicationRecordsNotificationTrigger";
import {MedicationRecordsTimeOfDayNotificationTrigger} from "./models/MedicationRecordsTimeOfDayNotificationTrigger";

export interface AppData {
    babies?: Baby[];
    babyDateOfBirth?: BabyDateOfBirth[];
    babyImages?: BabyImage[];
    excretionNotificationTriggers?: ExcretionNotificationTrigger[];
    pumpings?: Pumping[];
    pumpingNotificationTriggers?: PumpingNotificationTrigger[];
    pumpingTimeOfDayNotificationTriggers?: PumpingTimeOfDayNotificationTrigger[];
    sleeps?: Sleep[];
    sleepingNotificationTriggers?: SleepingNotificationTrigger[];
    teeths?: Teeth[];
    temperatures?: Temperature[];
    vaccinations?: Vaccination[];
    journalEntries?: JournalEntry[];
    feeds?: Feed[];
    excretions?: Excretion[];
    medicines?: Medicine[];
    pauses?: Pause[];
    growths?: Growth[];
    sleepingsTimeOfDayNotificationTriggers?: SleepingsTimeOfDayNotificationTrigger[];
    medicineRecords?: MedicineRecord[];
    excretionTimeOfDayNotificationTriggers?: ExcretionTimeOfDayNotificationTrigger[];
    feedingNotificationTriggers?: FeedingNotificationTrigger[];
    feedingTimeOfDayNotificationTriggers?: FeedingTimeOfDayNotificationTrigger[];
    generalNotesNotificationTriggers?: GeneralNotesNotificationTrigger[];
    generalNotesTimeOfDayNotificationTriggers?: GeneralNotesTimeOfDayNotificationTrigger[];
    medicationRecordsNotificationTriggers?: MedicationRecordsNotificationTrigger[];
    medicationRecordsTimeOfDayNotificationTriggers?: MedicationRecordsTimeOfDayNotificationTrigger[];
}